+++ plugins/catalog-backend-module-bitbucket-server/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @backstage/plugin-catalog-backend-module-bitbucket-server
 
+## 0.1.20-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-plugin-api@0.6.7-next.2
+  - @backstage/backend-common@0.19.9-next.2
+  - @backstage/backend-tasks@0.5.12-next.2
+  - @backstage/plugin-catalog-node@1.5.0-next.2
+
 ## 0.1.20-next.1
 
 ### Patch Changes