+++ plugins/home-react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @backstage/plugin-home-react
 
+## 0.1.5-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.13.8-next.2
+
 ## 0.1.5-next.1
 
 ### Patch Changes