+++ plugins/stack-overflow/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @backstage/plugin-stack-overflow
 
+## 0.1.22-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.13.8-next.2
+  - @backstage/frontend-plugin-api@0.3.0-next.2
+  - @backstage/plugin-home-react@0.1.5-next.2
+  - @backstage/plugin-search-react@1.7.2-next.2
+
 ## 0.1.22-next.1
 
 ### Patch Changes