+++ plugins/techdocs-addons-test-utils/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @backstage/plugin-techdocs-addons-test-utils
 
+## 1.0.23-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.13.8-next.2
+  - @backstage/plugin-catalog@1.15.0-next.2
+  - @backstage/plugin-techdocs@1.9.0-next.2
+  - @backstage/integration-react@1.1.21-next.1
+  - @backstage/plugin-search-react@1.7.2-next.2
+  - @backstage/plugin-techdocs-react@1.1.13-next.2
+
 ## 1.0.23-next.1
 
 ### Patch Changes