+++ plugins/user-settings/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @backstage/plugin-user-settings
 
+## 0.7.12-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.13.8-next.2
+  - @backstage/frontend-plugin-api@0.3.0-next.2
+  - @backstage/plugin-catalog-react@1.9.0-next.2
+
 ## 0.7.12-next.1
 
 ### Patch Changes